Update documenation to reflect changes
