+++ src/models/object_access_control.rs
@@ -35,8 +35,8 @@ pub struct ObjectAccessControl {
     pub generation: Option<String>,
     /// The entity holding the permission, in one of the following forms:
     ///
-    /// user-userId
-    /// user-email
+    /// # user-userId
+    /// # user-email
     /// group-groupId
     /// group-email
     /// domain-domain
